# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API de Gestão de Pessoas",
        "description": "Descrição da API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://project.test/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/pessoa": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Lista de pessoas",
                "description": "Retorna lista de pessoas",
                "operationId": "getPessoaList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PessoaCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/pessoa": {
            "post": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Cadastra nova pessoa",
                "description": "Retorna mensagem de sucesso",
                "operationId": "storePessoa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUpdatePessoaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pessoa cadastrada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messagem": {
                                            "type": "string",
                                            "example": "Pessoa cadastrada com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Erro ao tentar cadastrar pessoa.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messagem": {
                                            "type": "string",
                                            "example": "Não foi possível cadastrar a pessoa."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreUpdatePessoaRequest": {
                "title": "Store Project request",
                "description": "Store Project request body data",
                "required": [
                    "nome, sobrenome"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Fulano"
                    },
                    "sobrenome": {
                        "type": "string",
                        "example": "Souza"
                    },
                    "idade": {
                        "type": "string",
                        "example": "31"
                    },
                    "login": {
                        "type": "string",
                        "example": "fulano@gmail.com"
                    },
                    "senha": {
                        "type": "password",
                        "example": "123456"
                    },
                    "status": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "PessoaCollection": {
                "title": "PessoaResource",
                "description": "Pessoa resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PessoaResource"
                }
            },
            "PessoaResource": {
                "title": "PessoaResource",
                "description": "Pessoa resource",
                "properties": {
                    "": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PessoaResource"
                }
            },
            "Pessoa": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Pessoa",
            "description": "API Endpoints de Pessoas"
        }
    ]
}